backend/requirements.txt
# Core framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
pydantic>=2.5.0

# ASR Model (for Apple Silicon)
parakeet-mlx>=0.4.0

# Audio processing
numpy>=1.24.0
scipy>=1.11.0

# Hugging Face for model download
huggingface-hub>=0.20.0

# Utilities
python-multipart>=0.0.6
aiofiles>=23.2.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
